screenshot-taker/pyproject.toml
[project]
name = "screenshot-webpage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.3.0",
    "playwright>=1.53.0",
]
